pyproject.toml
[tool.poetry]
name = "empire-automation"
version = "0.1.0"
description = "AI agent automation system for Empire Builder's Textbook execution"
authors = ["Empire Builder"]
readme = "README.md"
packages = [{include = "empire_automation"}]

[tool.poetry.dependencies]
python = "^3.10"
crewai = "^0.1.0"
anthropic = "^0.18.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
celery = "^5.3.0"
redis = "^5.0.0"
sqlalchemy = "^2.0.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
streamlit = "^1.28.0"
langchain = "^0.1.0"
langchain-anthropic = "^0.1.0"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.1.0"
google-api-python-client = "^2.100.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

